settings/muted-topic: Move list-create code to designated file.
This code better fits in `settings_muted_topics`.
